+++ src/scale/mod.rs
@@ -4,10 +4,12 @@ mod band;
 mod colormap;
 mod format;
 mod linear;
+mod spec;
 mod ticks;
 mod time;
 
 pub use band::Band;
 pub use colormap::Colormap;
 pub use linear::Linear;
+pub use spec::Scale;
 pub use ticks::{Tick, Ticks};